make c2 compile
