+++ src/platforms/esp/32/clockless_block_esp32.h
@@ -4,7 +4,11 @@
 #include <stdint.h>
 #include "namespace.h"
 #include "clock_cycles.h"
+#include "esp_intr_alloc.h"
+#include "eorder.h"
+#include "transpose8x1_noinline.h"
 
+#define FASTLED_HAS_CLOCKLESS 1
 #define FASTLED_HAS_BLOCKLESS 1
 
 #define PORT_MASK (((1<<LANES)-1) & 0x0000FFFFL)
@@ -20,7 +24,19 @@ extern uint32_t _frame_cnt;
 extern uint32_t _retry_cnt;
 #endif
 
-template <uint8_t LANES, int FIRST_PIN, int T1, int T2, int T3, EOrder RGB_ORDER = GRB, int XTRA0 = 0, bool FLIP = false, int WAIT_TIME = 5>
+__attribute__((always_inline)) inline void interrupt_unlock() {
+	// ets_intr_unlock();
+	// TODO: imlement interrupt_unlock?
+	// These functions were mined out of the code below and
+	// made no-ops here. We probably want to implement them.
+}
+
+__attribute__((always_inline)) inline void interrupt_lock()  {
+	// ets_intr_lock();
+	// TODO: imlement interrupt_lock?
+}
+
+template <int FIRST_PIN, int T1, int T2, int T3, EOrder RGB_ORDER = GRB, int XTRA0 = 0, bool FLIP = false, int WAIT_TIME = 5, uint8_t LANES = 1>
 class InlineBlockClocklessController : public CPixelLEDController<RGB_ORDER, LANES, PORT_MASK> {
     typedef typename FastPin<FIRST_PIN>::port_ptr_t data_ptr_t;
     typedef typename FastPin<FIRST_PIN>::port_t data_t;
@@ -40,12 +56,14 @@ public:
 	/*uint32_t clocks = */
 	int cnt=FASTLED_INTERRUPT_RETRY_COUNT;
 	while(!showRGBInternal(pixels) && cnt--) {
-	    ets_intr_unlock();
+	    // ets_intr_unlock();
+		interrupt_unlock();
 #ifdef FASTLED_DEBUG_COUNT_FRAME_RETRIES
 	    ++_retry_cnt;
 #endif
 	    delayMicroseconds(WAIT_TIME * 10);
-	    ets_intr_lock();
+	    // ets_intr_lock();
+		interrupt_lock();
 	}
 	// #if FASTLED_ALLOW_INTTERUPTS == 0
 	// Adjust the timer
@@ -58,7 +76,7 @@ public:
 
     template<int PIN> static void initPin() {
 	if(PIN >= REAL_FIRST_PIN && PIN <= LAST_PIN) {
-	    _ESPPIN<PIN, 1<<(PIN & 0xFF)>::setOutput();
+	    _ESPPIN<PIN, 1<<(PIN & 0xFF), true>::setOutput();
 	    // FastPin<PIN>::setOutput();
 	}
     }
@@ -134,7 +152,8 @@ public:
 	}
 	allpixels.preStepFirstByteDithering();
 	
-	ets_intr_lock();
+	// ets_intr_lock();
+	interrupt_lock();
 	uint32_t _start = __clock_cycles();
 	uint32_t last_mark = _start;
 	
@@ -150,21 +169,24 @@ public:
 	    writeBits<8+XTRA0,0>(last_mark, b0, allpixels);
 	    
 #if (FASTLED_ALLOW_INTERRUPTS == 1)
-	    ets_intr_unlock();
+	    // ets_intr_unlock();
+		interrupt_unlock();
 #endif
 	    
 	    allpixels.stepDithering();
 	    
 #if (FASTLED_ALLOW_INTERRUPTS == 1)
-	    ets_intr_lock();
+	    // ets_intr_lock();
+		interrupt_lock();
 	    // if interrupts took longer than 45µs, punt on the current frame
 	    if((int32_t)(__clock_cycles()-last_mark) > 0) {
-		if((int32_t)(__clock_cycles()-last_mark) > (T1+T2+T3+((WAIT_TIME-INTERRUPT_THRESHOLD)*CLKS_PER_US))) { ets_intr_unlock(); return 0; }
+		if((int32_t)(__clock_cycles()-last_mark) > (T1+T2+T3+((WAIT_TIME-INTERRUPT_THRESHOLD)*CLKS_PER_US))) { interrupt_unlock(); return 0; }
 	    }
 #endif
 	};
 	
-	ets_intr_unlock();
+	// ets_intr_unlock();
+	interrupt_unlock();
 #ifdef FASTLED_DEBUG_COUNT_FRAME_RETRIES
 	++_frame_cnt;
 #endif
