+++ src/transpose8x1_noinline.h
@@ -1,0 +1,3 @@
+
+#pragma once
+void transpose8x1_noinline(unsigned char *A, unsigned char *B);
